Made spell casting functional! (See notes)
Only Fire1 works right now. All spell cast animations are placeholders.
To do later is to create more preparing anmiations and HUD.

Placed cap on spell levels and exp (3 and 16, respectively). There are
actually two caps; the redudancy is to prevent the game from breaking even
if a user tampers with variables in the debug console. (Or more likely, I
accidentally forget an edge case, causing the glitch.)